tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/033.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: innerHTML and scripts moved in DOM</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>
	<div></div>
	<script>log('inline script #1');
		// script added with innerHTML should not run..
		document.getElementsByTagName('div')[0].innerHTML = '<script>log("innerHTML script runs")<\/script><script src="scripts/include-1.js"><\/script>';
		try{
			document.body.appendChild( document.getElementsByTagName('div')[0].firstChild );
			document.body.appendChild( document.getElementsByTagName('div')[0].firstChild );
		}catch(e){
			log('ERROR while testing');
		}
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2']);
		t.done();
}
	onload = t.step_func(function(){setTimeout(t.step_func(test), 200);})
	</script>

</body></html>
